Update to latest version of PaperMod theme
I realized that even though I updated the submodule to point to my fork,
the origin remote was still pointing to the original repository.
I fixed by basically removing the submodule and adding it again this time
pointed to my for and added an upstream remote to the original repository.

The steps I followed were:

1. Remove the submodule:
   ```bash
   git submodule deinit -f themes/PaperMod
   git rm -f themes/PaperMod
   rm -rf .git/modules/themes/PaperMod
   ```
I had to add `-f` for force to those commands but I'm not entirely sure why git complained about that.

2. Add the submodule again:
    ```bash
    git submodule add https://github.com/bpugh/hugo-PaperMod.git PaperMod
    ```

3. Add the upstream remote in the submodule dirctory:
    ```bash
    git remote add upstream https://github.com/adityatelange/hugo-PaperMod.git
    ```

4. Pulled in the latest changes from upstream PaperMod
    ```bash
    git fetch upstream
    git merge upstream/master
    ```

5. Updated the submodule in the repo to point to latest
    ```bash
    git add themes/PaperMod
    git commit
    ```