Show only regular posts on home page
The papermod theme shows everything in `mainSections`
According to the hugo docs if they're not defined explicitly in the site
config then it picks the section with the most entries which will now be
the TILs.
So I just have to mainly set it to `post` in the config.

I was able to determine what it was by putting this in the template:
`sections: {{ site.Params.mainSections }}`